split controls into separate component
